pre-historic-legacy/03-shared/axiom-esports-data/api/requirements-dev.txt
# Development and Testing Requirements
# Install with: pip install -r requirements.txt -r requirements-dev.txt

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# HTTP client for testing
httpx>=0.25.0

# Linting and formatting
ruff>=0.1.0
black>=23.0.0
mypy>=1.7.0

# Type stubs
asyncpg-stubs>=0.29.0

# Development utilities
watchdog>=3.0.0
